requirements.txt
annotated-types
asgiref
Babel
certifi
channels
contourpy
crispy-bootstrap5
cycler
distlib
dj-database-url
Django
django-bootstrap-datepicker-plus
django-ckeditor
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
django-froala-editor
django-jazzmin
django-js-asset
django-money
django-tailwindcss
djangorestframework
env
filelock
fonttools
kiwisolver
matplotlib
mysqlclient
numpy
packaging
Pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
py-moneyed
pydantic
pydantic_core
pyparsing
python-dateutil
pytz
setuptools
six
sqlparse
typing_extensions
tzdata
virtualenv
wheel